pluggable line generation
