+++ Examples/DemoPlacementViewController.swift
@@ -102,11 +102,13 @@ class DemoPlacementViewController: UIViewController {
         for latlon in latlons {
             locations.append(CLLocation(latitude: latlon.0, longitude: latlon.1))
         }
-        guard let firstLocation = locations.first, let lastLocation = locations.last else { return }
-        terrainNode.drawPath(from: locations, cylinderRadius: 10, color: .orange, verticalOffset: 3)
-        let firstNode = terrainNode.projectedSphere(at: firstLocation, radius: 15, color: .yellow)
-        terrainNode.addChildNode(firstNode)
-        let lastNode = terrainNode.projectedSphere(at: lastLocation, radius: 20, color: .red)
-        terrainNode.addChildNode(lastNode)
+//        guard let firstLocation = locations.first, let lastLocation = locations.last else { return }
+//        terrainNode.drawPath(from: locations, cylinderRadius: 10, color: .orange, verticalOffset: 3)
+//        let firstNode = terrainNode.projectedSphere(at: firstLocation, radius: 15, color: .yellow)
+//        terrainNode.addChildNode(firstNode)
+//        let lastNode = terrainNode.projectedSphere(at: lastLocation, radius: 20, color: .red)
+//        terrainNode.addChildNode(lastNode)
+        
+        terrainNode.addPolyline(coordinates: locations, startRadius: 20, endRadius: 20, startColor: .red, endColor: .red)
     }
 }
